refactor(display): display 폴더를 만들고, 그 폴더에 리펙토링 한 코드를 집어 넣음
